Working OBJ Renderer, + more
Used HBM's NTM mod obj model renderer as a base and then configured it to work for my specific usage. There are multiple changes, but we now have converted the MWC Marine armor model and texture to be used in this 1.7.10 mod. The model still needs to be adjusted to sit correctly on the player model.

Currently creating a python gui program to help with the conversion of these obj's to work for the mod!